removed CareKitUI files, CareKit.SliderLogTaskView, and the slider task from the catalog
