+++ CareKit/CareKit/iOS/Task/SliderLogTaskViewModel.swift
@@ -1,0 +1,29 @@
+//
+//  SliderLogTaskView.swift
+//
+//  Created by Dylan Li on 5/26/20.
+//  Copyright © 2020 NetReconLab. All rights reserved.
+//
+#if !os(watchOS)
+
+import CareKitStore
+import CareKitUI
+import Foundation
+import SwiftUI
+
+public struct SliderLogTaskViewModel {
+    
+    /// The title text to display in the header.
+    public let title: String
+    
+    /// The detail text to display in the header.
+    public let detail: String?
+    
+    /// Instructions text to display under the header.
+    public let instructions: String?
+
+    /// Action to perform when the button is tapped.
+    public let action: (Double) -> Void
+}
+
+#endif
