+++ CareKit/CareKitTests/Task/TestSliderTaskView.swift
@@ -25,29 +25,10 @@ class TestSliderLogTaskView: XCTestCase {
 
     let eventQuery = OCKEventQuery(for: Date())
     let task = OCKTask(id: "", title: "", carePlanUUID: nil, schedule: .dailyAtTime(hour: 1, minutes: 0, start: Date(), end: nil, text: nil))
-    var staticView: CareKitUI.SliderLogTaskView<_SliderLogTaskViewHeader, _SliderLogTaskViewSlider> {
-        CareKitUI.SliderLogTaskView(title: Text(""),
-                                 detail: Text(""),
-                                 instructions: Text(""),
-                                 valuesArray: $valuesArray,
-                                 value: $value,
-                                 range: 0...10,
-                                 step: 2,
-                                 minimumImage: nil,
-                                 maximumImage: nil,
-                                 sliderStyle: .system,
-                                 action: { _ in })
-    }
 
     func testDefaultContentInitializers() {
         _ = CareKit.SliderLogTaskView(task: task, eventQuery: eventQuery, storeManager: controller.storeManager)
         _ = CareKit.SliderLogTaskView(taskID: "", eventQuery: eventQuery, storeManager: controller.storeManager)
         _ = CareKit.SliderLogTaskView(controller: controller)
-    }
-
-    func testCustomContentInitializers() {
-        _ = CareKit.SliderLogTaskView(task: task, eventQuery: eventQuery, storeManager: controller.storeManager) { controller, value, valuesArray in self.staticView }
-        _ = CareKit.SliderLogTaskView(taskID: "", eventQuery: eventQuery, storeManager: controller.storeManager) { controller, value, valuesArray  in self.staticView }
-        _ = CareKit.SliderLogTaskView(controller: controller) { controller, value, valuesArray  in self.staticView }
     }
 }
