+++ OCKCatalog/OCKCatalog/Sections/TaskSection.swift
@@ -163,75 +163,6 @@ private struct TaskView: View {
                                          detail: Text("Anytime"),
                                          state: .incomplete(Text("NO DATA")))
                 }
-            case .slider:
-                VStack(spacing: 16) {
-                    
-                    // System slider style, with images, with color gradient
-                    CareKit.SliderLogTaskView(taskID: OCKStore.Tasks.distress.rawValue,
-                                           eventQuery: .init(for: Date()), storeManager: storeManager) { controller, value, valuesArray in
-                        CareKitUI.SliderLogTaskView(title: Text((controller.viewModel?.title ?? "") + " (System style, images, gradient)"),
-                                                 instructions: controller.viewModel?.instructions.map(Text.init),
-                                                 valuesArray: valuesArray,
-                                                 value: value,
-                                                 range: 0...10,
-                                                 step: 1,
-                                                 minimumImage: Image(systemName: "chevron.down"),
-                                                 maximumImage: Image(systemName: "chevron.up"),
-                                                 minimumDescription: "None",
-                                                 maximumDescription: "A Lot",
-                                                 sliderStyle: .system,
-                                                 gradientColors: [.red, .yellow, .green],
-                                                 action: controller.viewModel?.action ?? { _ in })
-                    }
-                    
-                    // System slider style, no images
-                    CareKit.SliderLogTaskView(taskID: OCKStore.Tasks.distress.rawValue,
-                                           eventQuery: .init(for: Date()), storeManager: storeManager) { controller, value, valuesArray in
-                        CareKitUI.SliderLogTaskView(title: Text((controller.viewModel?.title ?? "") + " (System style)"),
-                                                 instructions: controller.viewModel?.instructions.map(Text.init),
-                                                 valuesArray: valuesArray,
-                                                 value: value,
-                                                 range: 0...10,
-                                                 step: 1,
-                                                 minimumDescription: "None",
-                                                 maximumDescription: "A Lot",
-                                                 sliderStyle: .system,
-                                                 action: controller.viewModel?.action ?? { _ in })
-                    }
-                    
-                    // Ticked slider style, with images, with color gradient
-                    CareKit.SliderLogTaskView(taskID: OCKStore.Tasks.distress.rawValue,
-                                           eventQuery: .init(for: Date()), storeManager: storeManager) { controller, value, valuesArray in
-                        CareKitUI.SliderLogTaskView(title: Text((controller.viewModel?.title ?? "") + " (Ticked style, images, gradient)"),
-                                                 instructions: controller.viewModel?.instructions.map(Text.init),
-                                                 valuesArray: valuesArray,
-                                                 value: value,
-                                                 range: 0...10,
-                                                 step: 1,
-                                                 minimumImage: Image(systemName: "chevron.down"),
-                                                 maximumImage: Image(systemName: "chevron.up"),
-                                                 minimumDescription: "None",
-                                                 maximumDescription: "A Lot",
-                                                 sliderStyle: .ticked,
-                                                 gradientColors: [.red, .yellow, .green],
-                                                 action: controller.viewModel?.action ?? { _ in })
-                    }
-                    
-                    // Ticked slider style, with images
-                    CareKit.SliderLogTaskView(taskID: OCKStore.Tasks.distress.rawValue,
-                                           eventQuery: .init(for: Date()), storeManager: storeManager) { controller, value, valuesArray in
-                        CareKitUI.SliderLogTaskView(title: Text((controller.viewModel?.title ?? "") + " (Ticked style)"),
-                                                 instructions: controller.viewModel?.instructions.map(Text.init),
-                                                 valuesArray: valuesArray,
-                                                 value: value,
-                                                 range: 0...10,
-                                                 step: 1,
-                                                 minimumDescription: "None",
-                                                 maximumDescription: "A Lot",
-                                                 sliderStyle: .ticked,
-                                                 action: controller.viewModel?.action ?? { _ in })
-                    }
-                }
             default:
                 EmptyView()
             }
